refactor: move zUf and zUfCode to a new misc module for better organization
