+++ src/dfe/nfe/layoutNfe/groupB.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 import { zCustom } from "@/utils/zCustom";
-import { zUfCode } from ".";
+import { zUfCode } from "./misc";
 import { schemaNfeNfRef } from "./groupBA";
 
 const invalidCnfs = new Set([
