+++ src/dfe/nfe/layoutNfe/index.ts
@@ -1,20 +1,6 @@
 import { XMLBuilder } from "fast-xml-parser";
 import { SignedXml } from "xml-crypto";
-import { z } from "zod";
-
-import { ufCodeMap } from "@/ufCode";
-import type { UF, UFCode } from "@/ufCode/types";
-
 import { schemaNfeInfNfe, type NfeInfNfe } from "./groupA";
-
-// NOTE: Hack pro zod aceitar tipos customizados (https://stackoverflow.com/a/73825370)
-const ufCodeList = Object.keys(ufCodeMap);
-const ufCodes: [UFCode, ...UFCode[]] = [ufCodeList[0], ...ufCodeList];
-export const zUfCode = () => z.enum(ufCodes);
-
-const ufList = Object.values(ufCodeMap);
-const ufs: [UF, ...UF[]] = [ufList[0], ...ufList];
-export const zUf = () => z.enum(ufs);
 
 /**
  * @description Representa as informações necessárias para assinar uma NFe.
